expands the map at any resolution!

+++ src/lab03/map_helper.py
@@ -149,7 +149,7 @@ def to_grid_cells(cells_to_paint, my_map):
     grid.cell_width = my_map.info.resolution
 
     grid.cells = []
-    # Fill in points
+
     for index2d in cells_to_paint:
         point = Point()
         point.x = index2d[0]
@@ -166,6 +166,6 @@ def index1d_to_index2d(index1d, my_map):
     return (x, y)
 
 def index2d_to_index1d(index2d, my_map):
-    return index2d[1] * my_map.info.width + index2d[0]
+    return int(index2d[1] * my_map.info.width + index2d[0])
